BE/inference-yolo/requirements.txt
opencv-python==4.8.1.78
ultralytics==8.0.196
torch==2.0.1
torchvision==0.15.2
paho-mqtt==1.6.1
mysql-connector-python==8.1.0
python-dotenv==1.0.0
numpy==1.24.3
Pillow==10.0.0
rtsp-client==0.2.0
aiortc==1.6.0
aiohttp==3.8.5
asyncio-mqtt==0.13.0
psutil==5.9.5
